kernel: fix 64 compiler warnings

+++ kernel/source/MemoryEditor.c
@@ -42,7 +42,7 @@ static void PrintMemoryLine(U32 Base) {
     STR Addr[16];
     STR Hexa[64];
     STR ASCII[64];
-    U8* Pointer = (U8*)Base;
+    U8* Pointer = (U8*)(UINT)Base;
     U8 Data = 0;
     U32 Index = 0;
